# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Daily Journal API",
    "version": "1.0.0",
    "description": "A RESTful API for managing daily journal entries with OAuth authentication"
  },
  "servers": [
    {
      "url": "https://cse341-rlcp.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
      "Journal": {
        "type": "object",
        "required": ["title", "content", "mood", "weather"],
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "description": "Title of the journal entry"
          },
          "content": {
            "type": "string",
            "maxLength": 5000,
            "description": "Content of the journal entry"
          },
          "mood": {
            "type": "string",
            "enum": ["very-happy", "happy", "neutral", "sad", "very-sad", "anxious", "excited", "grateful"],
            "description": "Current mood"
          },
          "weather": {
            "type": "string",
            "enum": ["sunny", "cloudy", "rainy", "snowy", "stormy", "foggy", "windy"],
            "description": "Current weather"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 50
            },
            "description": "Tags for the entry (optional)"
          },
          "isPrivate": {
            "type": "boolean",
            "description": "Whether the entry is private (optional)"
          },
          "location": {
            "type": "string",
            "maxLength": 100,
            "description": "Location (optional)"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "githubId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      }
    }
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/journals": {
      "get": {
        "summary": "Get all journals for authenticated user",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "List of journals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Journal"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new journal entry",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Journal"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Journal created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}